requirements.txt
streamlit
openai-whisper
setuptools-rust
google-generativeai
python-pptx
streamlit-audiorecorder

